requirements.txt
beautifulsoup4==4.11.2
certifi==2022.12.7
charset-normalizer==3.0.1
idna==3.4
PyYAML==6.0
requests==2.28.2
soupsieve==2.3.2.post1
urllib3==1.26.14
